fix: components/icon
