+++ frontend/src/lib/components/icons/home.svelte
