+++ frontend/src/lib/components/icons/notification.svelte
