+++ frontend/src/lib/components/icons/plus.svelte
