+++ frontend/src/lib/components/icons/search.svelte
