--- frontend/src/lib/icons/Popular.svelte
@@ -1,3 +1,0 @@
-<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M17 5V20M12 11V20M7 14V20" stroke="#9FB9D0" stroke-width="2" stroke-linecap="round"/>
-</svg>
